Refactor recompression into separate module
